Alunos respondendo questões discursivas

+++ database/migrations/2021_02_08_190115_add_instituicao_id_table_unidade_academicas.php
@@ -14,7 +14,7 @@ class AddInstituicaoIdTableUnidadeAcademicas extends Migration
     public function up()
     {
         Schema::table('unidade_academicas', function (Blueprint $table) {
-            $table->unsignedBigInteger('instituicao_id')->nullable();
+            $table->unsignedInteger('instituicao_id')->nullable();
             $table->foreign('instituicao_id')->references('id')->on('instituicoes');
         });
     }
