+++ database/migrations/2021_04_03_134401_add_instituicao_id_in_table_ciclos.php
@@ -14,7 +14,7 @@ class AddInstituicaoIdInTableCiclos extends Migration
     public function up()
     {
         Schema::table('ciclos', function (Blueprint $table) {
-            $table->unsignedBigInteger('instituicao_id')->nullable();
+            $table->unsignedInteger('instituicao_id')->nullable();
             $table->foreign('instituicao_id')->references('id')->on('instituicoes');
         });
     }
