+++ database/migrations/2021_06_03_025430_create_resposta_discursivas_table.php
@@ -16,9 +16,9 @@ class CreateRespostaDiscursivasTable extends Migration
         Schema::create('resposta_discursivas', function (Blueprint $table) {
             $table->increments('id');
             $table->longText('resposta_discursiva');
-            $table->unsignedBigInteger('aluno_id');
-            $table->unsignedBigInteger('simulado_id');
-            $table->unsignedBigInteger('questao_discursiva_id');
+            $table->unsignedInteger('aluno_id');
+            $table->unsignedInteger('simulado_id');
+            $table->unsignedInteger('questao_discursiva_id');
             $table->foreign('aluno_id')->references('id')->on('alunos');
             $table->foreign('simulado_id')->references('id')->on('simulados');
             $table->foreign('questao_discursiva_id')->references('id')->on('questao_discursivas');
